Another day, another commit

+++ word_button.gd
@@ -7,6 +7,7 @@ var bS:buttonStatus
 @export var buttonHandle:RichTextLabel
 @export var interactableColor:Color = Color.DARK_KHAKI
 @export var fadeCurve:Curve
+@export var actualTextHandle:RichTextLabel
 
 var wData:WordData
 
@@ -17,9 +18,17 @@ var commandItem:bool = false
 
 var active:bool = true
 
+@export var driftingCurve:Curve
+var driftingTween:SimonTween
+var driftOn:bool =false
+
+@export var shakeCurve:Curve
+var shakeTween:SimonTween
+
 # Called when the node enters the scene tree for the first time.
 func _ready():
 	pass # Replace with function body.
+	self_modulate.a = 0.0
 
 # Called every frame. 'delta' is the elapsed time since the previous frame.
 func _process(delta):
@@ -28,10 +37,33 @@ func _process(delta):
 func updateWord(isDiscoverable:bool = false):
 	buttonHandle.text = word
 
+func toggleDrifting(on:bool = true):
+	var r = randf_range(0.8,1.8)
+	if (!driftOn and on):
+		driftingTween = SimonTween.new()
+		driftingTween.createTween(buttonHandle,"position:y",3,1.25*r,driftingCurve,SimonTween.PINGPONG).setLoops(-1)
+	elif (driftOn and !on):
+		driftingTween.stop()
+		
+		buttonHandle.position = Vector2(0,0)
+	driftOn = on
+
+func fadeIn():
+	var r = randf_range(0.1,0.9)
+	var f = SimonTween.new()
+	self.modulate.a = 0
+	f.createTween(self,"modulate:a",1,r)
+
+func shake():
+	shakeTween = SimonTween.new()
+	await shakeTween.createTween(buttonHandle,"position:y",20,0.2,shakeCurve).tweenDone
+	
+	return true
 
 func updateWordText(w:String):
 	word = w
-	self.text = w
+	self.text = "[center]"+w
+	buttonHandle.text  = "[center]"+w
 
 func updateData(button:WordButton):
 	word = button.word
@@ -40,6 +72,7 @@ func updateData(button:WordButton):
 func disable():
 	buttonHandle.self_modulate = Color.WHITE
 	active = false
+	toggleDrifting(false)
 
 func SetInteractible():
 	bS = buttonStatus.TEXTLOG_INTERACTIBLE
@@ -55,15 +88,22 @@ func SetCommands():
 func SetNormal():
 	bS = buttonStatus.TEXTLOG_NORMAL
 
+func Die():
+	print("I'm dying now!")
+	var s = SimonTween.new()
+	await s.createTween(self,"modulate:a",-1,0.1).tweenDone
+	self.queue_free()
+
 func OnClicked():
 	if (Global.gamePaused):
 		return
 	print("Clicked on word: "+str(word)+" !!!")
 	match bS:
 		buttonStatus.TEXTLOG_INTERACTIBLE:
-			print("And it's discoverable??")
-			disable()
-			Global.InteractableWordClicked.emit(word,self)
+			if (active):
+				print("And it's discoverable??")
+				disable()
+				Global.InteractableWordClicked.emit(word,self)
 		buttonStatus.INVENTORY:
 			print("And its an inventory item")
 			Global.InventoryInteractibleClicked.emit(word,self)
@@ -72,15 +112,22 @@ func OnClicked():
 			Global.CommandInteractibleClicked.emit(word,self, true)
 	#self.queue_free()
 
-func moveButtonToLocation(goalButtonHandle:WordButton):
-	print("YEP")
+func moveButtonToLocation(goalObjectHandle:Node, deleteAtEnd:bool = true, fadeGoal:bool = true):
+	var origPos = self.global_position
 	var s = SimonTween.new()
 	var u = SimonTween.new()
-	var distanceDelta = goalButtonHandle.global_position - self.global_position + Vector2(0,self.size.y)
-	goalButtonHandle.modulate.a = 0
-	u.createTween(goalButtonHandle,"modulate:a",1,0.4,fadeCurve)
+	var distanceDelta = goalObjectHandle.global_position - self.global_position + Vector2(0,self.size.y)
+	if (fadeGoal):
+		goalObjectHandle.modulate.a = 0
+		u.createTween(goalObjectHandle,"modulate:a",1,0.4,fadeCurve)
 	await s.createTween(self,"global_position",distanceDelta,0.4).anotherParallel().createTween(self,"modulate:a",-1,0.4,fadeCurve).tweenDone
-	deleteObject()
+	
+	if (deleteAtEnd):
+		deleteObject()
+	else:
+		self.global_position = origPos
+		if (fadeGoal):
+			await s.createTween(self,"modulate:a",1,Global.veryShortPause).tweenDone
 
 func deleteObject():
 	self.queue_free()
